Started adding circleci
